pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["IlyaBulatau <ilyabulatau@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.101.1"}
aiohttp = "^3.8.5"
alembic = "^1.11.3"
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
pydantic-settings = "^2.0.3"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
